# Speed up RandomPolynomial by a factor 13