The first commit avoids having to do the multiplication twice and speeds up Pol.Irreducible:

```
name              old time/op  new time/op  delta
PolIrreducible-8  62.4ms ± 0%  57.4ms ± 0%  -8.11%  (p=0.000 n=10+10)
```

The second introduces a fast path for Pol.Mul(x, 2), for a combined speedup of:

    name                old time/op  new time/op  delta
    RandomPolynomial-8   102ms ±13%     8ms ± 8%  -92.55%  (p=0.000 n=10+10)
    PolIrreducible-8    62.4ms ± 0%   4.3ms ± 0%  -93.04%  (p=0.000 n=10+7)